User Address Schema models, and routers are created. Product model and User model now have a new field for address.

+++ routers/user_address_router.py
@@ -10,13 +10,19 @@ router = APIRouter(
 
 
 @router.post('/add', response_model=AddressPrivateDisplay)
-def add_address(request: AddressBase, db: Session = Depends(get_db), current_user: UserBase = Depends(get_current_user)):
+def add_address(request: AddressBase, db: Session = Depends(get_db),
+                current_user: UserBase = Depends(get_current_user)):
+    request.user_id = current_user.id
     return db_user_address.add_address(db, request)
 
 
-@router.get('/my_addresses', response_model=List[AddressPrivateDisplay])
-def my_addresses(id: int = Query(..., alias='user_id'), db: Session = Depends(get_db), current_user: UserBase = Depends(get_current_user)):
-    return db_user_address.get_address(db, id)
+@router.get('/my-addresses', response_model=List[AddressPrivateDisplay])
+def my_addresses(id: int = Query(..., alias='user_id'),
+                 db: Session = Depends(get_db),
+                 current_user: UserBase = Depends(get_current_user)):
+    if current_user.id != id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to see this address!")
+    return db_user_address.my_addresses(db, id)
 
 
 @router.get('', response_model=AddressPublicDisplay)
@@ -24,11 +30,30 @@ def show_address_publicly(id: int, db: Session = Depends(get_db)):
     return db_user_address.get_address(db, id)
 
 
-@router.put('/modify/{id}')
-def modify_address(id: int, request: AddressBase, db: Session = Depends(get_db), current_user: UserBase = Depends(get_current_user)):
+@router.get('/{id}', response_model=AddressPrivateDisplay)
+def my_default_address(id:int, db: Session = Depends(get_db),
+                       current_user: UserBase = Depends(get_current_user)):
+    address = db_user_address.get_address(db, id)
+    if current_user.id != address.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to see this address!")
+    return db_user_address.get_address(db, id)
+
+
+@router.put('/modify/{id}', response_model=AddressPrivateDisplay)
+def modify_address(request: AddressBase, id: int,
+                   db: Session = Depends(get_db),
+                   current_user: UserBase = Depends(get_current_user)):
+    address = db_user_address.get_address(db, id)
+    if current_user.id != address.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to modify this address!")
     return db_user_address.modify_address(db, id, request)
 
 
 @router.delete('/delete/{id}')
-def delete_address(id: int, db: Session = Depends(get_db), current_user: UserBase = Depends(get_current_user)):
+def delete_address(id: int,
+                   db: Session = Depends(get_db),
+                   current_user: UserBase = Depends(get_current_user)):
+    address = db_user_address.get_address(db, id)
+    if current_user.id != address.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this address!")
     return db_user_address.delete_address(db, id)
